Api Details are added in Database in settings.php Revinate Cron is removed and table is added in revinate-import.php View and update all fields are displaying

+++ revinate-import.php
@@ -5,7 +5,7 @@ Plugin URI: http://wordpress.org/
 Description: The best rating plugin for WordPress. Revinate Rating shows all multiple Hotel ratings through Revinate API
 Version: 1.0
 Author: Sakhatech
-Author URI: http://sakhatech.com
+Author URI: https://github.com/ratheeshpkr/revinateimport
 License: GPL2
 Text Domain: multi-rating
 Domain Path: languages
@@ -29,14 +29,19 @@ class Revinate {
 
 			global $wpdb; // Must have this or else!
 
-			$postmeta_table = $wpdb->postmeta;
+			 $postmeta_table = $wpdb->postmeta;
 			$posts_table = $wpdb->posts;
-
-			$wpdb->query("DELETE FROM " . $postmeta_table . " WHERE meta_key = 'link'");
-			//$wpdb->query("DELETE FROM " . $postmeta_table . " WHERE meta_key = '_mrlpt_client_phone_num'");
+			$option_table = $wpdb->options;
+			
+			$log_table = $wpdb->prefix . 'revinateLog';
+			//$wpdb->query("DELETE FROM " . $postmeta_table . " WHERE meta_key = 'link'");
+			$wpdb->query("DELETE FROM " . $postmeta_table . " WHERE meta_key IN('title','link','author','rating',
+				     'language','subratings','roomsubratings','valuesubratings','hotelsubratings',
+				     'locationsubratings','cleansubratings','triptype','pagesize','pagetotalele','pagetotalpage','numbers','authorlocation')");
 			$wpdb->query("DELETE FROM " . $posts_table . " WHERE post_type = 'revinate_reviews'");
-
-
+			$wpdb->query("DELETE FROM " . $option_table . " WHERE option_name IN('revin_settings_url','revin_settings_username','revin_settings_token','revin_settings_secret')");
+			$wpdb->query("TRUNCATE TABLE ".$log_table);########log need to be inserted for cron file
+			$wpdb->query("DROP TABLE ".$log_table);########log need to be inserted for cron file
 			flush_rewrite_rules();
 
 		}
@@ -55,10 +60,14 @@ class Revinate {
 		$db_version = '1.0';
 
 		global $wpdb;
-		
-		
-		require_once( ABSPATH . 'wp-admin/includes/upgrade.php' );
-		//dbDelta( $sql );
+		$table_name = $wpdb->prefix . 'revinateLog';
+		$sql = "CREATE TABLE $table_name (id int(11) NOT NULL AUTO_INCREMENT,
+		pageNo int(11) NOT NULL,
+		TotalPage int(11) NOT NULL,
+		Success int(11) NOT NULL,
+		UNIQUE KEY id (id))";
+		require_once( ABSPATH . 'wp-admin/includes/upgrade.php');
+		dbDelta( $sql );
 
 		add_option( 'db_version', $db_version );
 
@@ -69,100 +78,20 @@ class Revinate {
 	 *	Connection String for Revinate API
 	 *  Inserting Json Values from API
 	 */
-	function rev_install_data() {
-
- 		//$url = "https://porter.revinate.com/hotels/10463/reviews";
-		/*$url = "https://porter.revinate.com/hotels/10470";
-		$USERNAME="martin.rusteberg@snhgroup.com";
-		$TOKEN="ef74b36fe595cf9fdef0bce348616c3d";
-		$SECRET="f94c5129c8efd82a11c7a20c1471f77c4a08e922d9683b27456462e58878de19";   */
-		$url = get_option('revin_settings_url');
-		 
-		$USERNAME= get_option('revin_settings_username');
-		$TOKEN= get_option('revin_settings_token');
-		$SECRET= get_option('revin_settings_secret');    
-		$kSecret = crypt($SECRET,$const.substr(sha1(mt_rand()), 0, 22));
-		$TIMESTAMP = time();
-
-		$ENCODED = hash_hmac('sha256', $USERNAME.$TIMESTAMP,$SECRET);
-
-		$ch = curl_init();
-		curl_setopt($ch, CURLOPT_VERBOSE, TRUE);
-		curl_setopt($ch,CURLOPT_SSL_VERIFYPEER,0);
-		curl_setopt($ch, CURLOPT_HTTPHEADER, array(
-			'X-Revinate-Porter-Key:' .$TOKEN,
-			'X-Revinate-Porter-Username:' .$USERNAME,
-			'X-Revinate-Porter-Timestamp:' .$TIMESTAMP,
-			'X-Revinate-Porter-Encoded:' . $ENCODED,
-
-		));
-
-		curl_setopt($ch, CURLOPT_URL, $url);
-		curl_setopt($ch, CURLOPT_RETURNTRANSFER, 1);
-
-		$http_result = curl_exec($ch);
-		$error = curl_error($ch);
-
-		$http_code = curl_getinfo($ch );
-
-		curl_close($ch);
-		$arr =  json_decode($http_result,true);
-		$content = $arr['content'];
-		
-		global $wpdb;
-
-		// $table_name = $wpdb->prefix . 'revinate_reviews';
-		
-		foreach($content as $val){
-
-			$title = $val['title'];
-
-			if(!isset($title)){
-				$title = "";
-			}
-			
-			$querystr = "SELECT * FROM $wpdb->postmeta WHERE $wpdb->postmeta.meta_key = 'link' AND $wpdb->postmeta.meta_value = '".$val['links'][0]['href']."'";
-			$pageposts = $wpdb->get_results($querystr, OBJECT);
-			
-			if(count($pageposts) > 0){
-				continue;
-			}
-
-			$post_id = wp_insert_post(array (
-				'post_type' => 'revinate_reviews',
-				'post_title' => $val['title'],
-				'post_content' => $val['language']['englishName'],
-				'post_status' => 'publish',
-				'comment_status' => 'closed',   // if you prefer
-				'ping_status' => 'closed',      // if you prefer
-			));
-			
-			
-			if ($post_id) {
-				// insert post meta
-				add_post_meta($post_id, 'title', $val['title']);
-				add_post_meta($post_id, 'link', $val['links'][0]['href']);
-				add_post_meta($post_id, 'author', $val['author']);
-				add_post_meta($post_id, 'rating', $val['rating']);
-				add_post_meta($post_id, 'language', $val['language']['englishName']);
-				add_post_meta($post_id, 'subratings', $val['subratings']['Service']);
-				add_post_meta($post_id, 'roomsubratings', $val['subratings']['Rooms']);
-				add_post_meta($post_id, 'valuesubratings', $val['subratings']['Value']);
-				add_post_meta($post_id, 'hotelsubratings', $val['subratings']['Hotel condition']);
-				add_post_meta($post_id, 'locationsubratings', $val['subratings']['Location']);
-				add_post_meta($post_id, 'cleansubratings', $val['subratings']['Cleanliness']);
-				add_post_meta($post_id, 'triptype', $val['tripType']);
-				add_post_meta($post_id, 'pagesize', $val['page']['size']);
-				add_post_meta($post_id, 'pagetotalele', $val['page']['totalElements']);
-				add_post_meta($post_id, 'pagetotalpage', $val['page']['totalPages']);
-				add_post_meta($post_id, 'numbers', $val['page']['number']);
-			}
-			
-		}
-
-	}
-
+	
 }
+
+
+#############My code
+	function rev_install_data(){
+		
+	}
+	#############My code
+
+
+
+
+	
 	/**
 	 * Includes files
 	 */
@@ -218,6 +147,8 @@ class Revinate {
 
 		add_action( 'add_meta_boxes', 'add_reviews_metaboxes' );
 	}
+	
+	
 
 	// Add the Revinate Reviews Meta Boxes
 
@@ -235,9 +166,16 @@ class Revinate {
 			wp_create_nonce( plugin_basename(__FILE__) ) . '" />';
 
 		// Get the data if its already been entered
-			$title = get_post_meta($post->ID, 'title', true);
+	
+			/*if(!empty(get_post_meta($post->ID, 'title', true)))
+				$title = get_post_meta($post->ID, 'title', true);
+			else
+				$title = 'No Title';
+				*/
+			
 			$link = get_post_meta($post->ID, 'link', true);
 			$author = get_post_meta($post->ID, 'author', true);
+			$authorloc = get_post_meta($post->ID, 'authorlocation', true);
 			$language = get_post_meta($post->ID, 'language', true);
 			$rating = get_post_meta($post->ID, 'rating', true);
 			$subratings = get_post_meta($post->ID, 'subratings', true);
@@ -250,12 +188,14 @@ class Revinate {
 
 
 		// Echo out the field
-			echo '<p>Title:</p>';
-			echo '<input type="text" name="title" value="' . $title  . '" class="widefat" />';
+			//echo '<p>Title:</p>';
+			//echo '<input type="text" name="title" value="' . $title  . '" class="widefat" />';
 			echo '<p>Link</p>';
 			echo '<input type="text" name="link" value="' . $link  . '" class="widefat" />';
 			echo '<p>Author</p>';
 			echo '<input type="text" name="author" value="' . $author  . '" class="widefat" />';
+			echo '<p>Author Location</p>';
+			echo '<input type="text" name="author" value="' . $authorloc  . '" class="widefat" />';
 			echo '<p>Language</p>';
 			echo '<input type="text" name="language" value="' . $language  . '" class="widefat" />';
 			echo '<p>Rating</p>';
